+++ src/tutorials/11. Layout - Grid.ts
@@ -1,0 +1,16 @@
+export const TURORIAL_11_Grid = `
+
+<div class="grid grid-column-template(1fr,1fr)">
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+  <div class="w(50) h(50) b(#ccc) r(8) pack">1</div>
+</div>
+
+
+
+`.trim()